skni_site_backend/pyproject.toml
[project]
name = "sknisite"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "ruff (>=0.9.5,<0.10.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.3.1,<6.0.0)",
    "drf-yasg (>=1.21.7,<2.0.0)",
    "setuptools (>=68.0.0,<69.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
